Module Dhcp_client_lwt.Make

Parameters

module Net : Mirage_net.S

Signature

include Mirage_net.S

type error = private [>

| Mirage_net.Net.error

]

val pp_error : error Fmt.t

type t

val disconnect : t -> unit Lwt.t

val write : t -> size:int -> (Cstruct.t -> int) -> (unit, error) Stdlib.result Lwt.t

val listen : t -> header_size:int -> (Cstruct.t -> unit Lwt.t) -> (unit, error) Stdlib.result Lwt.t

val mac : t -> Macaddr.t

val mtu : t -> int

val get_stats_counters : t -> Mirage_net.stats

val reset_stats_counters : t -> unit

type lease = Dhcp_wire.pkt

val lease_mvar : t -> lease Lwt_mvar.t

val connect : ?renew:bool -> ?xid:Cstruct.uint32 -> ?options:Dhcp_wire.dhcp_option list -> ?requests:Dhcp_wire.option_code list -> Net.t -> t Lwt.t

connect ~renew ~xid ~options ~requests net starts a DHCP client communicating over the network interface net. The client will attempt to get a DHCP lease at least once, and will return any leases obtained in the stream returned by connect. If renew is true, which it is by default, the client will attempt to renew the lease according to the logic in RFC2131. If renew is false, the client will cancel its listener and end the stream once the first lease has been obtained. The options are the DHCP options sent by the client in the DHCP DISCOVER and DHCP REQUEST. The requests will be transmitted as DHCP parameter request also in the DHCP DISCOVER and DHCP REQUEST.

val connect_no_dhcp : Net.t -> t Lwt.t