Module Dhcp_ipv4

module type S = sig ... end

module type With_lease = sig ... end

module Make (Network : Mirage_net.S) (Ethernet : Ethernet.S) (Arp : Arp.S) : sig ... end

Make(N)(E)(A) instantiates a device that exposes a Mirage_net.S interface as well as a Tcpip.Ip.S interface with ipv4. On connection it either requests a DHCP lease or sets up a static IPv4 address. The exposed Mirage_net.S interface is N but with incoming dhcp messages filtered out. It is expected that E and A are instantiated with N.

module Proj_net (T : S) : sig ... end

Projection of the Mirage_net.S device.

module Proj_ipv4 (T : S) : sig ... end

Projection of the IPv4 Tcpip.Ip.S device.

module Proj_lease (T : With_lease) : sig ... end

Projection of the DHCP lease as a device.